+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## [Unreleased]
 ### Added
 - Autodetect editable diagrams.net files (#48)
+- Entries in create new file context menu for diagrams.net including an empty template for editable SVG, editable PNG and XML (#48)
 
 ### Changed
 - Update to Diagrams.net 14.6.0